[#1041] cmd/*: factor out autocomplet command generation
